pyproject.toml
[project]
name = "adzuna-job-market-elt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow>=3.1.8",
    "boto3>=1.42.74",
    "dbt-snowflake>=1.11.3",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "snowflake-connector-python>=4.3.0",
]
